requirements.txt
absl-py==2.1.0
asgiref==3.5.0
astunparse==1.6.3
Babel==2.8.0
bcrypt==3.2.0
blinker==1.4
# Brlapi==0.8.3
Brotli==1.0.9
certifi==2020.6.20
chardet==4.0.0
# chrome-gnome-shell==0.0.0
click==8.0.3
colorama==0.4.4
# command-not-found==0.3
cryptography==3.4.8
# cupshelpers==1.0
dbus-python==1.2.18
# defer==1.0.6
distro==1.7.0
fasteners==0.14.1
Flask==2.0.1
flatbuffers==25.2.10
future==0.18.2
gast==0.6.0
google-pasta==0.2.0
grpcio==1.71.0
# gyp==0.1
h11==0.13.0
h5py==3.13.0
httplib2==0.20.2
idna==3.3
importlib-metadata==4.6.4
itsdangerous==2.1.0
jeepney==0.7.1
Jinja2==3.0.3
keras==3.9.0
keyring==23.5.0
# language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
libclang==18.1.1
lockfile==0.12.2
# louis==3.20.0
macaroonbakery==1.3.1
Mako==1.1.3
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.0.1
mdurl==0.1.2
ml_dtypes==0.5.1
monotonic==1.6
more-itertools==8.10.0
mutagen==1.45.1
namex==0.0.8
netifaces==0.11.0
numpy==2.1.3
oauthlib==3.2.0
olefile==0.46
opt_einsum==3.4.0
optree==0.14.1
packaging==24.2
paramiko==2.9.3
pexpect==4.8.0
Pillow==9.0.1
protobuf==5.29.3
ptyprocess==0.7.0
pycairo==1.20.1
pycryptodomex==3.11.0
pycups==2.0.1
Pygments==2.19.1
PyGObject==3.42.1
pyinotify==0.9.6
PyJWT==2.3.0
pymacaroons==0.13.0
PyNaCl==1.5.0
pyOpenSSL==21.0.0
pyparsing==2.4.7
pyRFC3339==1.1
python-apt==2.4.0+ubuntu4
python-dateutil==2.8.1
python-debian==0.1.43+ubuntu1.1
pytz==2022.1
pyxattr==0.7.2
pyxdg==0.27
PyYAML==5.4.1
reportlab==3.6.8
requests==2.25.1
rich==13.9.4
SecretStorage==3.3.1
simplejson==3.17.6
six==1.16.0
systemd-python==234
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorflow-io-gcs-filesystem==0.37.1
tensorflow_cpu==2.19.0
termcolor==2.5.0
typing_extensions==4.12.2
ubuntu-drivers-common==0.0.0
ubuntu-pro-client==8001
ufw==0.36.1
unattended-upgrades==0.1
urllib3==1.26.5
usb-creator==0.3.7
uvicorn==0.15.0
wadllib==1.3.6
websockets==9.1
Werkzeug==2.0.2
wrapt==1.17.2
wsproto==1.0.0
xdg==5
xkit==0.0.0
yt-dlp==2022.4.8
zipp==1.0.0